Refactor metrics to be pull based.
This brings the collection model more in line with the official
Prometheus client guide. The types for metrics and descriptors have
been greatly simplified, descriptors are now factories for metrics,
metrics are simply collect functions along with a static handle value.

The primary benefit of the refactoring is that this will allow for
custom metrics that expose the current state of the system (e.g.
current CPU temp or number of threads) without the need for polling.