Flip a lot of types
Previously, a lot of metric-using functions had types such as

  observeDuration :: IO a -> Summary -> IO a

But in practice this is really awkward to use. Following the principle that the
"most variable" arguments should be at the end, we really want

  observeDuration :: Summary -> IO a -> IO a

As this allows us to write code such as:

  foo = observeDuration summary $ do
    ..
    ..

All metric manipulating functions have been modified to have the metric as the
first argument for consistency.