Use Text instead of String
Fixes #7.

The only complication is most users will now be using -XOverloadedStrings, which
is ambiguous when used with the current Label implementation. It could be argued
this was already a bug, as the same errors would have happened when we were
using strings.

To fix inference, I've made the decision that all of the tuple Label instances
can only be defined once, and they must be homogeneous tuples of Text. This
gives enough information to the type inference engine to avoid ambiguous types.