Fix observe with lazy metrics
observe for both Histograms and Summarys was previously throwing an exception
about `atomically` being nested. This seems to happen because the unsafeRegister
needs to use atomically to register the metric - but the metric is now only
being forced inside the observation itself.

Adding a bang pattern on the carrier for the metric seems to be sufficient. This
only needs to be done for metrics that use `atomically`, which is just Histogram
and Summary.